server breakdown
