+++ media/inject_script.html
@@ -5,7 +5,8 @@
 	const connection = new WebSocket(url);
 
 	connection.onerror = (error) => {
-		console.log('WebSocket error: ' + error);
+		console.log('WebSocket error: ');
+		console.log(error);
 	};
 
 	connection.onmessage = (e) => {
