+++ src/browserPreview.ts
@@ -104,9 +104,13 @@ export class BrowserPreview extends Disposable {
 		super.dispose();
 	}
 
+	private get _host() {
+		return `http://127.0.0.1:${this._port}`;
+	}
+
 	private handleOpenBrowser(givenURL: string) {
 		const urlString =
-			givenURL == '' ? this.constructHostAddress(this._panel.title) : givenURL;
+			givenURL == '' ? this.constructAddress(this._panel.title) : givenURL;
 		const url = vscode.Uri.parse(urlString);
 		vscode.env.openExternal(url);
 		vscode.window.showInformationMessage(
@@ -122,12 +126,13 @@ export class BrowserPreview extends Disposable {
 		}
 	}
 
-	private constructHostAddress(URLExt: string): string {
+	private constructAddress(URLExt: string): string {
 		if (URLExt.length > 0 && URLExt[0] == '/') {
 			URLExt = URLExt.substring(1);
 		}
-		return `http://127.0.0.1:${this._port}/${URLExt}`;
+		return `${this._host}/${URLExt}`;
 	}
+
 
 	private setHtml(webview: vscode.Webview, url: string): void {
 		this._panel.webview.html = this.getHtmlForWebview(webview, url);
@@ -179,7 +184,7 @@ export class BrowserPreview extends Disposable {
 				font-src ${this._panel.webview.cspSource};
 				style-src ${this._panel.webview.cspSource};
 				script-src 'nonce-${nonce}';
-				frame-src ${url};
+				frame-src ${this._host};
 				">
 				<meta name="viewport" content="width=device-width, initial-scale=1.0">
 
@@ -283,7 +288,7 @@ export class BrowserPreview extends Disposable {
 	}
 
 	private goToFile(URLExt: string): void {
-		this.setHtml(this._panel.webview, this.constructHostAddress(URLExt));
+		this.setHtml(this._panel.webview, this.constructAddress(URLExt));
 	}
 
 	private setPanelTitle(title = '/'): void {
