+++ src/manager.ts
@@ -1,6 +1,6 @@
 import { BrowserPreview } from './browserPreview';
 import { Disposable } from './dispose';
-import { Server } from './server';
+import { Server } from './server/serverManager';
 import * as vscode from 'vscode';
 import {
 	INIT_PANEL_TITLE,
@@ -63,7 +63,7 @@ export class Manager extends Disposable {
 
 		this.currentPanel.onDispose(() => {
 			this.currentPanel = undefined;
-			if (this._server.running) {
+			if (this._server.isRunning) {
 				vscode.window
 					.showInformationMessage(
 						'You closed the embedded preview. Would you like to also close the server?',
@@ -80,7 +80,7 @@ export class Manager extends Disposable {
 	}
 
 	public openServer(showMsgAlreadyOn = false): void {
-		if (!this._server.running) {
+		if (!this._server.isRunning) {
 			this._server.openServer(
 				this._serverPort,
 				this._serverWSPort,
@@ -93,7 +93,7 @@ export class Manager extends Disposable {
 	}
 
 	public closeServer(showMsgAlreadyOff = false): void {
-		if (this._server.running) {
+		if (this._server.isRunning) {
 			this._server.closeServer();
 
 			if (this.currentPanel) {
